esp32_device_service/requirements.txt
pyserial==3.5
Pillow==9.5.0
Flask==2.2.5
Flask-CORS==3.0.10
requests==2.31.0
